rename '___MemoryMap' traits to '___AddressBus'
and add 'set_address()' method to 'PpuAddressBus'. this method will be
needed for when the ppu address bus is updated even though no
reads/writes are happening (happens when 'current_vram_addr' is set).
the rename was needed to better match the new functionality that this
method adds (it makes no sense to call 'set_address()' on a memory map).