fix lifetime elision issue on raw trait object
- by adding a 'static bound on the 'PpuAddressBus' trait. this enforces
that implementors don't contain any non-'static references, and ensures
that all lifetime information is lost when casting to a raw pointer.
previously, the raw trait object 'ppu_bus_raw' in 'main()' would
secretly borrow 'bus', since the implicit 'a lifetime given to the trait
object itself would be elided and merged with the lifetime of 'bus' when
calling 'bus.base()'. I (stupidly) used a transmute to fix this, but
this is no longer needed.