store shared ref to framebuffer instead of raw ptr
now that 'PixelRenderer' provides an api that is entirely based on
shared references and internal mutability, it is perfectly legal to
simply store a shared reference to the framebuffer provided by
'PixelRenderer' (that allows mutation through Cell) for the entire
duration of the program. previously, a raw pointer was needed to get
around &mut exclusivity issues. though the use of a raw pointer was
safe (as far as i could tell, anyway - the pointer was guaranteed to
always be valid and didn't directly alias any other references to the
'PixelRenderer' struct), the safety was not 'codified' the way it is
now.